+++ mrcore/errors_storage.go
@@ -3,24 +3,24 @@ package mrcore
 import . "github.com/mondegor/go-sysmess/mrerr"
 
 var (
-	FactoryErrStorageConnectionIsAlreadyCreated = NewFactory(
+	FactoryErrStorageConnectionIsAlreadyCreated = NewFactoryWithCaller(
 		"errStorageConnectionIsAlreadyCreated", ErrorKindInternal, "connection '{{ .name }}' is already created")
 
-	FactoryErrStorageConnectionIsNotOpened = NewFactory(
+	FactoryErrStorageConnectionIsNotOpened = NewFactoryWithCaller(
 		"errStorageConnectionIsNotOpened", ErrorKindInternal, "connection '{{ .name }}' is not opened")
 
-	FactoryErrStorageConnectionFailed = NewFactory(
+	FactoryErrStorageConnectionFailed = NewFactoryWithCaller(
 		"errStorageConnectionFailed", ErrorKindSystem, "connection '{{ .name }}' is failed")
 
-	FactoryErrStorageQueryFailed = NewFactory(
+	FactoryErrStorageQueryFailed = NewFactoryWithCaller(
 		"errStorageQueryFailed", ErrorKindInternal, "query is failed")
 
-	FactoryErrStorageFetchDataFailed = NewFactory(
+	FactoryErrStorageFetchDataFailed = NewFactoryWithCaller(
 		"errStorageFetchDataFailed", ErrorKindInternal, "fetching data is failed")
 
 	FactoryErrStorageNoRowFound = NewFactory(
-		"errStorageNoRowFound", ErrorKindInternalNotice, "no row found")
+		"errStorageNoRowFound", ErrorKindInternal, "no row found")
 
 	FactoryErrStorageRowsNotAffected = NewFactory(
-		"errStorageRowsNotAffected", ErrorKindInternalNotice, "rows not affected")
+		"errStorageRowsNotAffected", ErrorKindInternal, "rows not affected")
 )