+++ mrserver/mrreq/errors_http.go
@@ -4,11 +4,11 @@ import . "github.com/mondegor/go-sysmess/mrerr"
 
 var (
 	FactoryErrHttpRequestCorrelationID = NewFactory(
-		"errHttpRequestCorrelationID", ErrorKindInternalNotice, "header 'X-Correlation-Id' contains incorrect value '{{ .value }}'")
+		"errHttpRequestCorrelationID", ErrorKindInternal, "header 'X-Correlation-Id' contains incorrect value '{{ .value }}'")
 
 	FactoryErrHttpRequestUserIP = NewFactory(
-		"errHttpRequestUserIP", ErrorKindInternalNotice, "UserIP '{{ .value }}' is not IP:port")
+		"errHttpRequestUserIP", ErrorKindInternal, "UserIP '{{ .value }}' is not IP:port")
 
 	FactoryErrHttpRequestParseUserIP = NewFactory(
-		"errHttpRequestParseUserIP", ErrorKindInternalNotice, "UserIP contains incorrect value '{{ .value }}'")
+		"errHttpRequestParseUserIP", ErrorKindInternal, "UserIP contains incorrect value '{{ .value }}'")
 )