+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # GoWebCore Changelog
 Все изменения библиотеки GoWebCore будут документироваться на этой странице.
 
+## 2024-03-18
+### Changed
+- Внедрена новая версия библиотеки `go-sysmess`, в связи с этим:
+  - всем ошибкам с типами `ErrorKindInternal` и `ErrorKindSystem` конструктор заменён на `NewFactoryWithCaller()`;
+  - константа `ErrorKindInternalNotice` переименована в `ErrorKindInternal`;
+  - централизовано логируются ошибки, у которых `appError.HasCallStack() == true`;
+- В `mrzerolog.CallerWithSkipFrame` добавлена логика отключающая автоматический вызов `CallStack()`;
+- Поправлены `skipFrame` в некоторых вызовах `Caller()`;
+
 ## 2024-03-17
 ### Changed
 - Для `mrparser.ItemStatus` добавлена возможность указания значения по умолчанию,